The Den Markings Extras (#1395)
* Hooks Wings to Unathi ; Plus Other Unathi and Vulpkanin Markings (#692)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This re-enables the markings layer for unathi's wing layer, but also
includes a couple of markings as a preparation for the next ones.
This does also add the inner overlay for dragon wing added in
SimpleStation resource folders, as well as the harpy one that puts them
over arms. (Note to cherrypick: remove harpy if no likey)

---

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Reconnect and reuse wings marking layer for unathi.
- [x] Add a couple of vulpkanin markings (Neck-chest fluff and Brushtip
tail)
- [x] Added cheesewedge markings on vulpkanins. (merp not included YET)
- [x] Added Sharp Cobra Hood

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

These are all examples! You don't have to do them like it..

![image](https://github.com/user-attachments/assets/7499fc48-c3c6-46ad-ab02-fd54cbd5ed65)

![image](https://github.com/user-attachments/assets/9417b2d4-b5e0-4691-b369-f64899a147d8)

![image](https://github.com/user-attachments/assets/9ddb15bd-741e-4676-be5a-1a504a599842)

![image](https://github.com/user-attachments/assets/0416612d-27d4-4829-8994-68d64f3eea37)
![image](https://github.com/user-attachments/assets/db9811d1-03ca-4b6e-a755-94c912356c2d)

</p>
</details>

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added cheesedreamer markings for vulps... but no ᵐᵉʳᵖ ⁽ʸᵉᵗ!⁾
- add: A couple of vulpkanin markings and an unathi marking.
- tweak: Wings now able to be attached (decoratively) to Unathi. (NT
forbids their use for flight, HOWEVER.)

---------

Co-authored-by: chly <haychiethewonderfulsnivy@gmail.com>

* fixes

* species restrictions

* hello

* Test

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Test 2

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Test 3

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Test 4

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Test 5

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Test 6

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Test 7

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

* Apply suggestions from code review

* Update Resources/Locale/en-US/_DEN/markings/reptilianandvulpkanin.ftl

Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>

---------

Co-authored-by: LaconiumDragona <riascazadofelra@gmail.com>
Co-authored-by: chly <haychiethewonderfulsnivy@gmail.com>
Co-authored-by: TheRealMasterChief117 <200016362+TheRealMasterChief117@users.noreply.github.com>
Co-authored-by: M3739 <47579354+M3739@users.noreply.github.com>