Big Food Port (#1245)
* passive mob changes

* Big Food Cherry-Pick (#351)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A big merge of a bunch of different PRs, listed below

https://github.com/new-frontiers-14/frontier-station-14/pull/2282
https://github.com/space-wizards/space-station-14/pull/24799
https://github.com/new-frontiers-14/frontier-station-14/pull/1823
https://github.com/new-frontiers-14/frontier-station-14/pull/2047
https://github.com/new-frontiers-14/frontier-station-14/pull/2023
https://github.com/space-wizards/space-station-14/pull/30116

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Pears
- [x] Avocado and Cucumbers
- [x] Honey
- [x] Bagels
- [x] Spicy Pizza
- [x] Raisins

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media (credit to the original PRs)</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/385f73ac-06b1-40b3-8d1b-a77ad994a115)

![image](https://github.com/user-attachments/assets/8f65747e-8cea-4f7e-be22-900ffacf1537)

![image](https://github.com/user-attachments/assets/7cd932b5-8e7a-414e-b40f-692e9c5e50de)

![image](https://github.com/user-attachments/assets/09e84746-bc9d-4fbe-8354-3c90402bbcb1)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Various Frontier and Wizden Contributors
- add: Pears and Pear related foods
- add: Avocados and related foods
- add: Cucumbers and their related foods
- add: Honey and their related foods
- add: Uranium Pizza, it's hot, it's crunchy, it's spicy, but most of
all, IT'S FRESH OUT THE OVEN.
- add: Other various food items (including but not limited too: Bagels)

* Raisins Now Contain Raisins (#529)

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

I dunno what else to tell ya, they got raisins in them now

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: 4no Raisins now contain raisins in them (for recipe purposes)

* raisin sprite now work

* bark

* expensive seeds

* bees cost more money

* requested fixes, meow

---------

Co-authored-by: Rosycup <178287475+Rosycup@users.noreply.github.com>